requirements.txt
fastapi==0.115.11
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
SQLAlchemy==2.0.38
